# scr/dokuSwagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ÜK M295 LB",
    "version": "3.0.5",
    "description": "API zur Verwaltung von Aufgaben"
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "Benutzeranmeldung",
        "tags": ["Authentifizierung"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Erfolgreiche Anmeldung",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Ungültige Anmeldeinformationen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fehler"
                }
              }
            }
          }
        }
      }
    },
    "/verify": {
      "get": {
        "summary": "Benutzerangemeldet überprüfen",
        "tags": ["Authentifizierung"],
        "responses": {
          "200": {
            "description": "Benutzer ist angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fehler"
                }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "delete": {
        "summary": "Benutzerabmeldung",
        "tags": ["Authentifizierung"],
        "responses": {
          "204": {
            "description": "Erfolgreich abgemeldet"
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fehler"
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Alle Aufgaben abrufen",
        "tags": ["Aufgaben"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Liste der Aufgaben",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Aufgabe"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fehler"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Neue Aufgabe erstellen",
        "tags": ["Aufgaben"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Aufgabe"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Aufgabe erfolgreich erstellt",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Aufgabe"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fehler"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Aufgabe anhand der ID abrufen",
        "tags": ["Aufgaben"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aufgabe gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Aufgabe"
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fehler"
                }
              }
            }
          },
          "404": {
            "description": "Aufgabe nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fehler"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Aufgabe anhand der ID aktualisieren",
        "tags": ["Aufgaben"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Aufgabe"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Aufgabe erfolgreich aktualisiert",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Aufgabe"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fehler"
                }
              }
            }
          },
          "404": {
            "description": "Aufgabe nicht gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fehler"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Aufgabe anhand der ID löschen",
        "tags": ["Aufgaben"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }]
          }}}}
          